phase 5: preset system + handoff doc
Source/PresetManager.{h,cpp} (new):
- 16 factory presets in kFactoryPresets[]:
  * 6 from Python BUILTIN_PRESETS: Psytrance (Default), Techno Deep,
    Hardstyle Zap, Projektor Punch, Projektor Deep Sub, Projektor Hard F#
  * 10 from research/04 §7: Forest Stomp 145, Hard Psy Main F#, Deep Twilight G,
    Progressive 138, Hi-Tech Snap A, Darkpsy Stomp E, Astrix Main F#,
    Burn In Noise Forest, Sub Lover 142, Tight Club F#
- applyParameterMap() converts raw values -> normalized via RangedAudioParameter
  range, then setValueNotifyingHost for proper DAW automation feedback
- saveKickPreset / loadKickPreset: APVTS XML, the DAW-portable format
- saveJson / loadJson: flat dict, Python-compatible
  * 'invert_phase' written as JSON bool to match Python convention
  * loadJson tolerant of unknown keys (KickAss-only params absent in Python
    presets simply leave APVTS at default)
- User preset dir resolved via juce::File::getSpecialLocation (portability rule
  §8b#2 — works on Windows %APPDATA% AND macOS ~/Library/Application Support/)
- rescanUserPresets() picks up both .kickpreset and .json files

Source/PluginEditor.{h,cpp}:
- Editor now implements juce::FileDragAndDropTarget — drag .json or .kickpreset
  onto window, red border highlight during drag, loads on drop
- New header controls: preset combo (factory + user with separators),
  note-snap combo (Off, C1..C2), SAVE button, LOAD button
- populatePresetCombo() builds the dropdown with --- Factory --- / --- User ---
  separators (disabled items)
- handleNoteSnapSelection() sets end_freq to chosen note's Hz via APVTS
- doSavePreset / doLoadPreset use async juce::FileChooser; member persists
  to survive the lambda lifetime (SnareRhythmGen pattern)

Source/PluginProcessor.{h,cpp}:
- Forward declare PresetManager, owned via std::unique_ptr (so we can include
  PresetManager.h only in .cpp)
- Explicit destructor declared in header + defined as default in .cpp (required
  with forward-declared unique_ptr type)
- getPresetManager() accessor for the editor

Source/WaveformDisplay.cpp:
- Fix: juce::jlimit/jmax template deduction fails with std::atomic<float>&
  argument despite implicit conversion — added explicit (float) casts on the
  4 affected lines (pitch_curve, vol_sustain, vol_curve, scoop_depth)

CMakeLists.txt: added PresetManager.cpp to sources.

docs/HANDOFF.md (new):
- Complete next-session entry point with: what's working, what needs manual
  testing (9 specific checklists), known issues, Phase 6 priority list,
  v1.x backlog, architecture cheat-sheet, files map, build instructions.

Build verified clean: VST3 8.11 MB, Standalone 8.12 MB, both fresh at 23:58.

+++ Source/WaveformDisplay.cpp
@@ -129,18 +129,18 @@ void WaveformDisplay::recomputeEnvelopeTraces()
     const float endF    = *processor.apvts.getRawParameterValue ("end_freq");
     const float t1Ms    = *processor.apvts.getRawParameterValue ("sweep_time_1");
     const float t2Ms    = *processor.apvts.getRawParameterValue ("sweep_time_2");
-    const float pCurve  = juce::jmax (0.01f, *processor.apvts.getRawParameterValue ("pitch_curve"));
+    const float pCurve  = juce::jmax (0.01f, (float) *processor.apvts.getRawParameterValue ("pitch_curve"));
 
     const float ta  = *processor.apvts.getRawParameterValue ("vol_attack");
     const float th  = *processor.apvts.getRawParameterValue ("vol_hold");
     const float td1 = *processor.apvts.getRawParameterValue ("vol_decay_1");
-    const float sus = juce::jlimit (0.0f, 100.0f, *processor.apvts.getRawParameterValue ("vol_sustain")) * 0.01f;
+    const float sus = juce::jlimit (0.0f, 100.0f, (float) *processor.apvts.getRawParameterValue ("vol_sustain")) * 0.01f;
     const float td2 = *processor.apvts.getRawParameterValue ("vol_decay_2");
-    const float vC  = juce::jmax (0.01f, *processor.apvts.getRawParameterValue ("vol_curve"));
+    const float vC  = juce::jmax (0.01f, (float) *processor.apvts.getRawParameterValue ("vol_curve"));
 
     const float scStart  = *processor.apvts.getRawParameterValue ("scoop_start");
     const float scLength = *processor.apvts.getRawParameterValue ("scoop_length");
-    const float scDepth  = juce::jlimit (0.0f, 100.0f, *processor.apvts.getRawParameterValue ("scoop_depth")) * 0.01f;
+    const float scDepth  = juce::jlimit (0.0f, 100.0f, (float) *processor.apvts.getRawParameterValue ("scoop_depth")) * 0.01f;
 
     const int  width = juce::jmax (1, getWidth() - (int) (kLeftAxisPx + 4.0f));
     ampEnvTrace.assign ((size_t) width, 0.0f);
